lm-watermarking/requirements.txt
gradio
nltk
scipy
torch
transformers
tokenizers
django
djangorestframework
python-dotenv
accelerate
django-cors-headers